@@ -1,5 +1,5 @@
 module github.com/neucn/neugo
 
-go 1.14
+go 1.16
 
 require github.com/stretchr/testify v1.6.1
